Use decimals instead of floats so Rails uses BigDecimal for calculations.

+++ db/migrate/20111227150106_create_orders.rb
@@ -8,9 +8,9 @@ class CreateOrders < ActiveRecord::Migration
       t.string :email, :null => false
       t.string :phone, :null => false
 
-      t.float :total, :null => false
-      t.float :total_due, :null => false
-      t.float :tax_charge, :null => false
+      t.decimal :total, :null => false
+      t.decimal :total_due, :null => false
+      t.decimal :tax_charge, :null => false
       t.string :status, :null => false
 
       t.timestamps
@@ -19,3 +19,4 @@ class CreateOrders < ActiveRecord::Migration
 end
 
 # Note: To force precision, alter column type in database console
+# Always use decimal when dealing with currency
